Refactoring Qr code verification & move all qr code into components/qrcode
