+++ components/qrcode/qr-generator.jsx
