+++ components/qrcode/qr-reader.jsx
