+++ pages/org/quotient/loan.jsx
@@ -12,7 +12,7 @@ import FormFieldCarDetails from 'components/forms/newLoan/form-field-car-details
 import FormFieldNameAndBirthday from 'components/forms/form-field-id';
 import FormFieldAddress from 'components/forms/form-field-address';
 import FormFieldPersonalContact from 'components/forms/form-field-personal-contact';
-import QrCodeAuthentication from 'components/qr-auth';
+import QrCodeAuthentication from 'components/qrcode/qr-auth';
 import QuotientSuccess from 'components/org/quotient/quotient-success';
 import { PROOFT_TEMPLATES_IDS } from 'utils/constants';
 import qrCodeVerificationData from 'data/qrcode-text-data';
